tests/test-requirements.txt
requests==2.24.0
mysql-connector-python==8.0.21
selenium==3.141.0
PyYAML==5.3.1
boto3==1.15.5
hvac==0.10.5
python-dotenv==0.14.0
pytest==6.0.2
